State switching complete, example started, bonuses
StateBot now manages which SimpleState is running, and it has methods to easily switch the current_state. SimpleStates are easily created and edited in the editor, and they also contain all necessary functions and overridable methods needed for state processing and switching. I started work on an example using a simple up/down/wait counter. It currently counts up and then deactivates the StateBot. The "bonuses.gd" script currently has a get_all_children() function, which is very useful.